requirements.txt
agno
gemini
gitpython
pinecone-client
requests